Extend Entity class and add simple AI
Added a simple ai demo where a second entity follows the player. This entity can jump up stairs or over gaps. Currently this entity has an issue where it can jump with 2x the standard force. A fix is being worked on, but it involves reworking the Entity.jump() system.